Fix unnecessary type cast in phase drift card
Removed redundant 'as int' type cast in PhaseDriftCard._buildJournalCountCard()
method. The count variable is already typed as int from the map values.

This improves code clarity and follows Dart best practices for type safety.

+++ lib/features/insights/cards/phase_drift_card.dart
@@ -246,7 +246,7 @@ class PhaseDriftCard extends StatelessWidget {
 
     int totalCount = 0;
     for (final count in counts.values) {
-      totalCount += count as int;
+      totalCount += count;
     }
     final dateStr = _formatDate(timestamp);
